pyproject.toml
[project]
name = "meeting-summary"
version = "0.1.0"
description = "Watch local call recordings, transcribe with faster-whisper, and summarize with Ollama."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "faster-whisper>=1.1.1",
  "python-dotenv>=1.1.0",
  "requests>=2.32.3",
  "watchdog>=6.0.0",
]

[project.scripts]
meeting-summary = "meeting_summary.main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["meeting_summary*"]

